rootfs/requirements.txt
# Drycc passport requirements
django==2.2.18
django-auth-ldap==2.2.0
django-cors-middleware==1.5.0
django-guardian==2.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
jsonfield==3.1.0
psycopg2-binary==2.8.5
requests==2.24.0
requests-toolbelt==0.9.1
django_redis==4.12.1
dj-database-url==0.5.0
django-oauth-toolkit==1.4.1
